requirements.txt
aiogram==3.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
asyncpg==0.29.0